+++ packages/kernel/src/rect.ts
@@ -21,3 +21,48 @@ export function rectContains(container: Rect, nested: Rect): boolean {
 export function rectArea(rect: Rect): number {
   return rect.width * rect.height;
 }
+
+/** Point-in-rect with inclusive edges on all four bounds. */
+export function containsPoint(rect: Rect, x: number, y: number): boolean {
+  return x >= rect.x && x <= rect.x + rect.width && y >= rect.y && y <= rect.y + rect.height;
+}
+
+export function pickLargestRect(rects: Rect[]): Rect | null {
+  let best: Rect | null = null;
+  let bestArea = -1;
+  for (const rect of rects) {
+    const area = rect.width * rect.height;
+    if (area > bestArea) {
+      best = rect;
+      bestArea = area;
+    }
+  }
+  return best;
+}
+
+/** Inclusive-edge overlap on both axes: touching a viewport edge still counts as visible. */
+export function isRectVisibleInViewport(targetRect: Rect, viewportRect: Rect): boolean {
+  return (
+    rangesOverlapInclusive(
+      targetRect.x,
+      targetRect.x + targetRect.width,
+      viewportRect.x,
+      viewportRect.x + viewportRect.width,
+    ) &&
+    rangesOverlapInclusive(
+      targetRect.y,
+      targetRect.y + targetRect.height,
+      viewportRect.y,
+      viewportRect.y + viewportRect.height,
+    )
+  );
+}
+
+function rangesOverlapInclusive(
+  leftStart: number,
+  leftEnd: number,
+  rightStart: number,
+  rightEnd: number,
+): boolean {
+  return Math.max(leftStart, rightStart) <= Math.min(leftEnd, rightEnd);
+}
