+++ scripts/layering/facade-symbols.ts
@@ -602,18 +602,15 @@ export const FACADE_SYMBOLS: readonly (readonly [string, readonly string[]])[] =
     '@agent-device/contracts/snapshot',
     [
       'buildSnapshotNodeMap',
-      'containsPoint',
       'extractNodeText',
       'findNearestScrollableAncestor',
       'isFillableType',
       'isNodeVisibleInEffectiveViewport',
       'isNodeVisibleOnScreen',
-      'isRectVisibleInViewport',
       'isScrollableNodeLike',
       'isScrollableType',
       'isTapPointInsideViewport',
       'normalizeType',
-      'pickLargestRect',
       'resolveEffectiveViewportRect',
       'resolveViewportRect',
     ],
@@ -786,7 +783,17 @@ export const FACADE_SYMBOLS: readonly (readonly [string, readonly string[]])[] =
     ],
   ],
   ['@agent-device/kernel/collections', ['uniqueStrings']],
-  ['@agent-device/kernel/rect', ['isPositiveFiniteRect', 'rectArea', 'rectContains']],
+  [
+    '@agent-device/kernel/rect',
+    [
+      'containsPoint',
+      'isPositiveFiniteRect',
+      'isRectVisibleInViewport',
+      'pickLargestRect',
+      'rectArea',
+      'rectContains',
+    ],
+  ],
   ['@agent-device/kernel/redaction', ['redactDiagnosticData']],
   ['@agent-device/kernel/bounds', ['parseBounds']],
   [
